refactor: 공용 테스트 컴포넌트 추상화
