+++ src/components/molecules/cardModal/cardModal.stories.jsx
@@ -1,4 +1,4 @@
-import { FormProvider, useForm } from "react-hook-form";
+import Form from "../../../utils/test/form";
 import CardModal from ".";
 
 export default {
@@ -7,15 +7,6 @@ export default {
   parameters: {
     layout: "fullscreen",
   },
-};
-
-const Form = (props) => {
-  const { children } = props;
-  const form = useForm({
-    defaultValues: {},
-  });
-
-  return <FormProvider {...form}>{children}</FormProvider>;
 };
 
 const FormTemplate = (args) => (
