+++ src/components/molecules/inputLength/inputLength.test.jsx
@@ -1,19 +1,8 @@
-import { FormProvider, useForm } from "react-hook-form";
 import { afterEach, describe, expect, it } from "vitest";
 import PropTypes from "prop-types";
 import { cleanup, render, screen } from "@testing-library/react";
+import Form from "../../../utils/test/form";
 import InputLength from ".";
-
-const Form = (props) => {
-  const { children } = props;
-  const form = useForm({
-    defaultValues: {
-      name: "test",
-    },
-  });
-
-  return <FormProvider {...form}>{children}</FormProvider>;
-};
 
 Form.propTypes = {
   children: PropTypes.node,
