+++ src/utils/storybook/form.jsx
@@ -1,0 +1,8 @@
+const Form = (props) => {
+  const { children } = props;
+  const form = useForm();
+
+  return <FormProvider {...form}>{children}</FormProvider>;
+};
+
+export default Form;
